+++ docs/decorator/script.js.deps
@@ -1,8 +1,9 @@
 file:///C:/tools/dart-sdk/lib/_internal/dart2js_platform.dill
 file:///C:/tools/dart-sdk/lib/libraries.json
-file:///D:/GitHub/patterns/docs/decorator/boardgames.dart
-file:///D:/GitHub/patterns/docs/decorator/chess.dart
-file:///D:/GitHub/patterns/docs/decorator/decorators.dart
+file:///D:/GitHub/patterns/docs/decorator/boarddecorators.dart
+file:///D:/GitHub/patterns/docs/decorator/game.dart
+file:///D:/GitHub/patterns/docs/decorator/gameboard.dart
+file:///D:/GitHub/patterns/docs/decorator/pieces.dart
 file:///D:/GitHub/patterns/docs/decorator/script.dart
 file:///D:/GitHub/patterns/docs/decorator/strategy.dart
 org-dartlang-sdk:///lib/_http/crypto.dart
