+++ docs/decorator/strategy.dart
@@ -1,59 +1,5 @@
-import 'boardgames.dart';
-import 'chess.dart';
-
-class ChessPiece extends GamePiece {
-  int i = 0;
-  int j = 0;
-  MovementStrategy moveStrategy;
-  String colour;
-  String name;
-  bool selected = false;
-  bool hasMoved = false;
-  int initialRow = -1;
-
-  ChessPiece(this.colour, this.name, this.moveStrategy) {
-    setSource("./assets/chess/${name}_${colour}.png");
-  }
-
-  void setPosition(int iInput, int jInput) {
-    i = iInput;
-    j = jInput;
-  }
-
-  List<MoveOption> move(ChessBoard board) {
-    List<MoveOption> options = moveStrategy.move(board, this);
-    print("the piece has ${options.length} options");
-    return options;
-  }
-
-  bool canCapture(ChessBoard board, int i, int j) {
-    if (validCoords(i, j)) {
-      GamePiece? target = board.getPiece(i, j);
-      if (target is ChessPiece && target.colour != colour) {
-        board.addMarker(i, j, "circle");
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  bool canMove(ChessBoard board, int i, int j) {
-    if (validCoords(i, j)) {
-      GamePiece? target = board.getPiece(i, j);
-      if (target == null) {
-        board.addMarker(i, j, "dot");
-        return true;
-      }
-    }
-
-    return false;
-  }
-
-  bool validCoords(int i, int j) {
-    return (0 <= i && i < 8) & (0 <= j && j < 8);
-  }
-}
+import 'gameboard.dart';
+import 'pieces.dart';
 
 class MoveOption {
   int i;
